+++ src/components/Pages/PagesStyled.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const Svg = styled.img`
+  height: 100%
+  width: 10x0%;
+`;
+export const Words = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
